Travels_App_DjangRF-main/Travels_App_DjangRF-main/requirements.txt
Django==4.2
# >>>>>>> REPLACE
djangorestframework==3.15.1
django-cors-headers==4.3.1
razorpay
